ejercicios resueltos

+++ ejercicio5/src/main/java/ar/edu/unlp/info/oo1/ejercicio5/Cuadrado.java
@@ -2,25 +2,25 @@ package ar.edu.unlp.info.oo1.ejercicio5;
 
 public class Cuadrado extends Figura2D{
 	
-	private float lado;
+	private double lado;
 	
 	public Cuadrado() {
 		this.lado = 0;
 	}
 	
-	public void setLado(float valor) {
+	public void setLado(double valor) {
 		this.lado = valor;
 	}
 	
-	public float getLado() {
+	public double getLado() {
 		return this.lado;
 	}
 	
-	public float getPerimetro() {
+	public double getPerimetro() {
 		return 4 * this.lado;
 	}
 	
-	public float getArea() {
+	public double getArea() {
 		return this.lado * this.lado;
 	}
